+++ cookbooks/arcgis-pro/metadata.rb
@@ -4,9 +4,9 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs and configures ArcGIS Pro'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '3.5.0'
+version          '3.6.0'
 
-depends          'arcgis-repository', '~> 3.5'
+depends          'arcgis-repository', '~> 3.6'
 depends          'windows', '~> 5.3'
 
 supports         'windows'